requirements.txt
google-cloud-bigquery
google-auth-oauthlib